requirements.txt
google-generativeai
IPython
pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
kokoro>=0.9.2
soundfile
numpy
